Disable RedirectHttp10Compatible to stay on https (#113)
Move VelocityViewResolver out of WebMvcConfigurer in order to give VelocityViewResolver a higher order. 

If not, the implicit MustacheViewResolver will be used before VelocityViewResolver during the redirect flow (return "redirect:/ ...") which will redirect back to Http even if Scoold is hosted on Https. And a redirect during a POST call will fail (Such like delete your question).